Added "strategy use" scoring script
This script is for scoring strategy use data. The subjects saw a description of each strategy type (1-7) and then typed in a box which ones they used. The script takes their response and creates 7 new variables that hold a 1 if they reported using that response. This data can then be used for further analysis on rates of strategy usage and success rates.